[Kernels][GPU] Small cleanups to the matmul kernels, NFC (#44855)
Uses keyword arguments, UInt, etc. to cleanup the gemv/gevm/matmul
implementations.

KERNEL_BENCHMARKS_ORIG_REV_ID: b16d836f6504b9f4e32a277c7c62d1a24ab45334

+++ mojo/kernels/benchmarks/gpu/bench_gemv_bf16.mojo
@@ -70,9 +70,9 @@ fn bench_gemv_tc[
                 c_buf.buffer.ptr,
                 a_buf.buffer.ptr,
                 b_buf.buffer.ptr,
-                M,
-                N,
-                K,
+                UInt(M),
+                UInt(N),
+                UInt(K),
                 grid_dim=ceildiv(M, WARPS_PER_BLOCK),
                 block_dim=WARP_SIZE * WARPS_PER_BLOCK,
             )
